% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{fix_path}
\alias{fix_path}
\title{A little fix for file.path}
\usage{
fix_path(path)
}
\description{
Ok, file.path don't check if the path already have a '/' at the start or end of parameters, so you can have double "/" ...
}
