% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdfs.R
\name{hadoop_command}
\alias{hadoop_command}
\title{Get the absolute path to hadoop bin directory}
\usage{
hadoop_command(command = "")
}
\arguments{
\item{command}{The hadoop command to get the absolute path of}
}
\value{
The absolute path of the Hadoop command
}
\description{
Get the bin path of Hadoop, HADOOP_HOME and HADOOP_BIN must be set in your Renviron.site
}
