% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdfs.R
\name{hdfs_command}
\alias{hdfs_command}
\title{Get the absolute path to 'hdfs' command}
\usage{
hdfs_command(options = "")
}
\arguments{
\item{options}{Option to git to the hdfs command}
}
\value{
The absolute path of the hdfs command
}
\description{
Get the absolute path of the hdfs command
}
