% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdfs.R
\name{hdfs_dfs_command}
\alias{hdfs_dfs_command}
\title{Get the absolute path to 'hdfs dfs' command}
\usage{
hdfs_dfs_command(options)
}
\arguments{
\item{options}{Option to git to the dfs command}
}
\value{
The absolute path of the dfs command
}
\description{
Get the absolute path of the hdfs command
}
