% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdfs.R
\name{hdfs_dfs_ls}
\alias{hdfs_dfs_ls}
\alias{hdfs_ls}
\title{Liste file/directory on HDFS}
\usage{
hdfs_ls(options = "")
}
\arguments{
\item{options}{Options to git to ls}
}
\value{
List of files/directories with their size
}
\description{
Liste file on hadoop, get file size
}
