% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{loadWorkspaceHDFS}
\alias{loadWorkspaceHDFS}
\title{Load the workspace from HDFS}
\usage{
loadWorkspaceHDFS(filename = NULL)
}
\arguments{
\item{filename}{The filename to load on HDFS. If null, filname will be getwd() + "/.RData".}
}
\value{
boolean
}
\description{
Load Workspace from HDFS
}
