% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdfs.R
\name{readHDFSFile}
\alias{readHDFSFile}
\title{Read a file from HDFS}
\usage{
readHDFSFile(filename, options = "rb")
}
\arguments{
\item{filename}{The filename on HDFS to load}

\item{options}{File mode}
}
\value{
A file object that you can use with load
}
\description{
Get a file on HDFS
}
