% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{saveWorkspaceHDFS}
\alias{saveWorkspaceHDFS}
\title{Save Workspace on HDFS}
\usage{
saveWorkspaceHDFS(filename = NULL)
}
\arguments{
\item{filename}{Path to the file on HDFS. If null, filname will be getwd() + "/.RData".}
}
\value{
boolean
}
\description{
Save your workspace on HDFS
}
