% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdfs.R
\name{writeHDFSFile}
\alias{writeHDFSFile}
\title{Write a file on HDFS}
\usage{
writeHDFSFile(filename, options = "wb")
}
\arguments{
\item{filename}{The filename on HDFS to save}

\item{options}{File mode}
}
\value{
A file object that you can use with save
}
\description{
Save a file on HDFS
}
